Started implementing timers

+++ firmware/src/common.cc
@@ -2,7 +2,7 @@
 
 void delay_ms ( int ms )
 {
-	volatile uint64_t nCount= ( CRYSTAL/10000/2 ) *ms;
+	volatile int nCount= ( CRYSTAL/10000/2 ) *ms;
 	for ( ; nCount!=0; nCount-- );
 }
